Agregue componente DatosPersonales, cosas de estilos
