+++ clientnative/App.js
@@ -6,8 +6,8 @@ import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
 import SingIn from "./Components/SingIn";
 import SingUp from "./Components/SingUp";
-import EditProfile from "./Components/EditProfile";
-import EditProfileCarrier from "./Components/EditProfileCarrier";
+import EditProfile from "./Components/Editares/EditProfile";
+import EditProfileCarrier from "./Components/Editares/EditProfileCarrier";
 import CompleteProfileUser from "./Components/CompletarPerfil/CompleteProfileUser";
 import CompleteProfileCarrier from "./Components/CompletarPerfil/CompleteProfileCarrier";
 import ProfileUserScreen from "./Components/ProfileUserScreen";
@@ -20,6 +20,7 @@ import StartCarrier from "./Components/ComenzarViaje/StartCarrier";
 import ScreenWaiting from "./Components/ScreenWaiting";
 import PaymentApp from "./Components/Payment/PaymentApp";
 import StartUser from "./Components/ComenzarViaje/StartUser";
+import EditVehicule from "./Components/Editares/EditVehicule";
 
 const Stack = createStackNavigator();
 
@@ -43,6 +44,11 @@ const App = () => {
             <Stack.Screen
               name="EditProfileCarrier"
               component={EditProfileCarrier}
+              options={{ headerShown: false }}
+            />
+             <Stack.Screen
+              name="EditVehicule"
+              component={EditVehicule}
               options={{ headerShown: false }}
             />
 
@@ -91,7 +97,6 @@ const App = () => {
               component={DatosPersonalesUser}
               options={{ headerShown: false }}
             />
-
             <Stack.Screen
               name="StartUser"
               component={StartUser}
