--- clientnative/Components/Config.js
@@ -1,3 +1,0 @@
-export const API_URL = 'http://192.168.0.105:3001'
- // esta es mi direccion, cambien por la suya
- // 'http://localhost:3001' --> o probar este si estamos desde la web
